Move the SecToTime function
From the models/issue_stopwatch.go file to the modules/util package
